+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - **select-option**: add a documentation about handling large text in options
+- **grid**: add support for row styling with css parts
 
 ### Changed